FEAT-1:Change cmake paths to match Cyclica chemnix docker file
We were using different Python library paths in the cmake options here
and in the Cyclica chemnix docker file. Although both sets of paths
were symlinked to the same files in Ubuntu Linux 18.04, it is less
confusing if we use the same paths in both installers.

See here for the Cyclica chemnix docker file:
https://github.com/cyclica/Dockerfile/blob/master/chemnix/Dockerfile