FEAT-1: Switch to --postgresql-root option to do Postgres cart build
The previous option did not specify the PostgreSQL installation root.
That was to be specified in an undocumented environment variable. The
PostgreSQL RDKit cartridge build may still work without that environment
variable set, but it’s better to set it.

To have the installer build the PostgreSQL cartridge now you now specify
the --postgresql-root option and give it the root of your PostgresSQL
isntallation. For PostgreSQL 10 on Ubuntu that would be
/usr/lib/postgresql/10. You no longer need to set an environment
variable. The installer passes the PostgreSQL root location using the
cmake flag instead.

It also prints out instructions on how to install the PostgreSQL
cartridge once the RDKit installation is complete.