BUG-4: Ask for Postgres version and install only its packages
If you are using the PostgreSQL community apt sources and you install
postgresql-server-dev-all, RDKit will find dev files for all available
Postgres versions and it will not know which one to build with. You have
to install only the server dev package for your version to keep it from
mismatching.