FEAT-2: Set default_python_version based on Ubuntu version
It also exits with an error if a non supported Ubuntu version is used.

It currently only supports Ubuntu 18.04 and 20.04. It skips the non
major releases because it has not been tested on them. You could get it
to work by disabling this check and specifying the correct Python and
boost versions for any Ubuntu release, but that is a bit beyond what
most users would be comfortable doing.