FEAT-2: Only exit on unsupported Ubuntu version when no --python-version
It will only exit with an error on an unsupported Ubuntu version if you
do not specify a --python-version. The error message tells you to
specify a python-version.

I think I will change this to default the Python version to 3. I think
that will work in all the package names where we currently use the point
version (3.6, 3.8, etc.).