FEAT-2: Have checksum patch apply to all RDKit versions
Searching for the changed font file checksum and updating it is harmless
if the old checksum is not in the file. Running the sed command to
update it whatever the RDKit version will fix it wherever it shows up
and allow us to install any of the versions that were broken by this
external dependency changing.